Rename file of update-schedules mutation hook
